requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
jinja2==3.1.4
python-dotenv==1.0.1
redis==5.0.7
httpx==0.27.0
pytest==8.2.0
